Splitting message converter into header converter and default message id strategy
